pod-engine/requirements.txt
flask==3.0.0
flask-cors==4.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
python-dotenv==1.0.0
rembg==2.0.50
pillow==10.2.0
onnxruntime==1.16.3
tenacity==8.2.3
pydantic==2.5.3
websockets==12.0
pandas==2.1.4
openpyxl==3.1.2
twilio==8.11.0
sendgrid==6.11.0
stripe==7.9.0
boto3==1.34.27
reportlab==4.0.9
